mysql-to-Datawarehouse/shop-analytics-level3/requirements.txt
pandas
sqlalchemy
pymysql
google-cloud-bigquery
pyarrow
python-dotenv